% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot_hierarchical.R
\name{barplot_hierarchical}
\alias{barplot_hierarchical}
\title{Hierarchical Composition Barplot with Dendrogram}
\usage{
barplot_hierarchical(
  ps,
  taxrank,
  feature,
  top,
  colors_barplot,
  colors_annotation,
  dist,
  size,
  decreasing
)
}
\arguments{
\item{ps}{A phyloseq object containing OTU table and taxonomy table}

\item{taxrank}{Taxonomic rank to use for coloring barplot (e.g., "Genus")}

\item{feature}{Metadata feature to annotate in color strips}

\item{top}{Number of top taxa to show individually}

\item{colors_barplot}{Color palette for taxonomic groups}

\item{colors_annotation}{List of color palettes for each annotation feature}

\item{dist}{Distance method for clustering ("bray", "aitchinson")}

\item{size}{Relative heights for plot components
(dendrogram, annotation, barplot)}

\item{decreasing}{Logical for ordering "other" category position}
}
\value{
A patchwork composite ggplot object
}
\description{
Creates a stacked barplot of taxonomic compositions
with hierarchical clustering dendrogram and sample metadata annotations.
}
\examples{
\dontrun{
data(GlobalPatterns)
tax_cols <- colnames(phyloseq::tax_table(GlobalPatterns))
other_tax <- tax_table_other(tax_cols)

barplot_hierarchical(
  ps = GlobalPatterns,
  taxrank = "Genus",
  feature = "SampleType",
  top = 10,
  colors_barplot = c("steelblue", "darkorange"),
  colors_line = c("red", "blue"),
  dist = "bray",
  size = c(1, 0.2, 2),
  decreasing = TRUE
)
}

}
