% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_prevalence.R
\name{compare_prevalence_across_group}
\alias{compare_prevalence_across_group}
\title{Generate Prevalence Comparison Plots for All Pairwise Group Combinations}
\usage{
compare_prevalence_across_group(
  physeq,
  column,
  point_color = "#d5b3ff",
  smooth_color = "#d62828",
  text_size = 3,
  zoom_plot_flag = T,
  zoom_window = list(x = c(0.1, 0.5), y = c(0.1, 0.5))
)
}
\arguments{
\item{physeq}{A phyloseq object}

\item{column}{Name of the column in sample_data containing groups to compare}

\item{point_color}{Color for scatter plot points (default: "#d5b3ff")}

\item{smooth_color}{Color for LOESS smooth line (default: "#d62828")}

\item{text_size}{Size for text labels (default: 3)}

\item{zoom_plot_flag}{add zoom for top taxa}

\item{zoom_window}{list of x and y range (example : list(x = c(0.1, 0.5), y = c(0.1, 0.5)))}

\item{top_n}{Number of top taxa to label in zoom plots (default: 10)}
}
\value{
A list of ggplot objects for each pairwise group comparison
}
\description{
Generate Prevalence Comparison Plots for All Pairwise Group Combinations
}
