% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_palette}
\alias{get_palette}
\title{Get color palette by name}
\usage{
get_palette(name)
}
\arguments{
\item{name}{Palette name (e.g., "set_2_1")}
}
\value{
Vector of HEX color codes
}
\description{
Get color palette by name
}
\examples{
get_palette("set_2_1")
}
