% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha.R
\name{plot_alpha_div}
\alias{plot_alpha_div}
\title{Plot Alpha Diversity (Basic)}
\usage{
plot_alpha_div(ps_obj, group, color, measure)
}
\arguments{
\item{ps_obj}{A phyloseq object containing microbiome data}

\item{measure}{Alpha diversity measure to plot (default: 'Shannon')}

\item{col}{Character string specifying the metadata column to group by}
}
\description{
Internal function for basic alpha diversity plot
}
