% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_table.R
\name{prepare_phyloseq}
\alias{prepare_phyloseq}
\title{Prepare Phyloseq Object with Transformations and Core Taxa Filtering}
\usage{
prepare_phyloseq(
  phyloseq,
  transformation = "compositional",
  detection = 0,
  prevalence = 0.1,
  taxonomic_level = "Species",
  first = "transform",
  filter_column = NULL,
  filter_value = NULL,
  filter_unclassified = T
)
}
\arguments{
\item{phyloseq}{A phyloseq object to process.}

\item{transformation}{Transformation method:
'compositional' (relative abundance), 'clr' (centered log-ratio) or NONE.}

\item{detection}{Abundance threshold for considering
taxa presence}

\item{prevalence}{Prevalence threshold
(proportion of samples where taxa must be present; 0-1).}

\item{taxonomic_level}{Taxonomic level for aggregation (e.g., "Genus").}

\item{first}{Parameter indicating that the data
is filtered or transformed the first.}

\item{filter_column}{Column name for sample filtering (optional)}

\item{filter_value}{Value to filter by in filter_column (optional)}

\item{filter_unclassified}{is flag to filter Kingdom == UNCLASSIFIED from MetaPhlAn}
}
\value{
A processed phyloseq object with core taxa meeting criteria.
}
\description{
Applies taxonomic aggregation, transformation
and core taxa filtering to a phyloseq object.
}
\examples{
# ps <- prepare_table(phyloseq_obj, transformation = "compositional",
#                    detection = 0.001, prevalence = 0.1,
# taxonomic_level = "Genus", first = 'filter')
}
