% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_plot_fonts}
\alias{setup_plot_fonts}
\title{Setup Fonts for Plotting}
\usage{
setup_plot_fonts(font_family = "Times New Roman")
}
\arguments{
\item{font_family}{Font family name to load (default: "Times New Roman")}
}
\description{
Initialize and load Times New Roman font for ggplot2
}
