% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_palette}
\alias{show_palette}
\title{Display color palette}
\usage{
show_palette(name)
}
\arguments{
\item{name}{Palette name}
}
\description{
Display color palette
}
\examples{
show_palette("set_10_1")
}
